backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Config
pydantic-settings==2.6.1
pydantic[email]==2.10.3

# Utilities
python-dotenv==1.0.1
